+++ web/Controllers/EditorsController.cs
@@ -69,6 +69,7 @@ namespace ASC.Api.Web.Help.Controllers
                 "ConversionApi",
                 "DemoPreview",
                 "DocumentBuilderApi",
+                "Drupal",
                 "Example/Java",
                 "Example/JavaSpring",
                 "Example/Nodejs",
@@ -99,6 +100,7 @@ namespace ASC.Api.Web.Help.Controllers
                 "Mattermost",
                 "Mentions",
                 "Methods",
+                "MobileIntegration",
                 "Moodle",
                 "Nextcloud",
                 "Nuxeo",
@@ -133,6 +135,7 @@ namespace ASC.Api.Web.Help.Controllers
                 "WOPI/RestApi/RefreshLock",
                 "WOPI/RestApi/RenameFile",
                 "WOPI/RestApi/Unlock",
+                "Wordpress"
             };
 
         [ValidateInput(false)]
@@ -285,6 +288,11 @@ namespace ASC.Api.Web.Help.Controllers
             return View();
         }
 
+        public ActionResult Drupal()
+        {
+            return View();
+        }
+
         public ActionResult Editor()
         {
             return View();
@@ -331,6 +339,11 @@ namespace ASC.Api.Web.Help.Controllers
         }
 
         public ActionResult Methods()
+        {
+            return View();
+        }
+
+        public ActionResult MobileIntegration()
         {
             return View();
         }
@@ -431,6 +444,11 @@ namespace ASC.Api.Web.Help.Controllers
                 catchall = null;
             }
             return View("WOPI", (object)catchall);
+        }
+
+        public ActionResult Wordpress()
+        {
+            return View();
         }
     }
 }