+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+- editors: added the Drupal ONLYOFFICE connector module
+- editors: added the Mobile integration page
+- editors: added the WordPress plugin
+
 ## 2.8.1
 - docbuilder: fix run examples